[WIP] Draft resources to obtain weka backend ip addresses
A new script is used to wait for the cluster to clusterize. Once
clusterized, we then get the details on the instance group for the weka
backends. From the weka backends, we pull data on each of the instances
and obtain a list of the backend ip addresses.

The network_storage output is populated with the backend IP address for
the first weka backend.